+++ src/util/util_bit.h
@@ -30,6 +30,7 @@
 #include "util_likely.h"
 #include "util_math.h"
 
+#include <cstddef>
 #include <cstdint>
 #include <cstring>
 #include <iterator>
